requirements.txt
geopandas==0.14.1
osmnx==1.7.1
pandas==2.1.4
numpy==1.26.2
shapely==2.0.2
folium==0.15.1
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
pyproj==3.6.1
rtree==1.1.0